requirements.txt
Flask
langchain
openai
python-dotenv
langchain-community
pypdf
langchain-openai
langchain-chroma
